Split out procedures files
